search README.md for broken links
tmcw/notfoundbot only searches in the configured `content-folder` - but
README.md is in the root directory, hence it was never searched.

This now upgrades tmcw/notfoundbot from 2.1.0 to 2.2.0, which is the
current version and required to use the `content-folder` input, and
configures it to search the entire root folder of the repo.